+++ CadEditor/Video.cs
@@ -297,7 +297,7 @@ namespace CadEditor
             }
 
             var bitmap = new Bitmap(ConfigScript.getScreenWidth()*64, ConfigScript.getScreenHeight()*64); //getScreenVertical, scales
-            byte[] indexes = Globals.getScreen(scrNo);
+            int[] indexes = Globals.getScreen(scrNo);
             using (var g = Graphics.FromImage(bitmap))
             {
                 for (int tileNo = 0; tileNo < SCREEN_SIZE; tileNo++)
